ARM: dts: omap5: Add aliases for rproc nodes
commit f6a856da4f3b10864135626ea8e4be4f86d50bfc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add aliases for the DSP and IPU remoteproc processor
nodes common to all OMAP5 boards. The aliases uses
the stem "rproc", and are identical to the values
chosen on OMAP4 boards.

The aliases can be overridden, if needed, in the
respective board files.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>