ARM: dts: dra72x: Add aliases for rproc nodes
commit 582d0394e1a7e6bc7a25ce188ee92f3ee12d6c36 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add aliases for all the 3 remote processor nodes common to
all DRA72x/DRA71x/AM571x/AM570x boards. The aliases uses the
stem "rproc", and are defined in the order of the most common
processors on the DRA72x family. The ids are same as DRA74x
except for the missing DSP2.

The aliases can be overridden, if needed, in the respective
derivative board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>