Introduce a new dtsi layer for S32V234EVB board
commit 9dd1f03615f789d51f8b58b78562f87a2097cc70 from
https://source.codeaurora.org/external/autobsps32/linux

S32V234 EVB board can be configured to use the Ethernet through BCM
switch or daughter PHY board. Due to the exclusivity of these two
configurations, a new dtsi layer is introduced to keep basic
configuration for the board IPs.The new dts files override the
fec node with the specific configurations.

Signed-off-by: Stoica Cosmin-Stefan <cosminstefan.stoica@freescale.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>