STM driver support
commit ebee200468b3408204c5f29c217a59b55a84400f from
https://source.codeaurora.org/external/autobsps32/linux

The driver can be used any two cores.
STM is not suitable for using it as a clocksource because the
counter is shared between channels so it is shared with the
channel used for clockevents. To avoid losing events, the counter is
stopped in the event interrupt, leading to inaccurate values for the
clocksource.

Issue: ALB-420

Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>