ARM: dts: keystone-k2l: Reserve SRAM for MPM
commit b7023a418d9a64c1326ddcf626c4e43e35bf868f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a child SRAM node to reserve a portion of the Multicore Shared
Memory (MSM) RAM for use by the keystone-dsp-mem driver on 66AK2L
SoCs. This memory will be exposed to the userspace for meeting the
needs of the Multi Proc Manager (MPM) stack.

A preliminary size of 512 KB is reserved to begin with and can be
adjusted as per need.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>