clk: Update the s32v234 clock tree for gating/ungating
commit ed7f3c417da73e6fbef66e82f1d4b380758058c6 from
https://source.codeaurora.org/external/autobsps32/linux

The s32v234 clock tree is updated for gating/ungating only
for the drivers builtin with the exception of ENET, because it's
clock schema it is not clear for the moment.

Using MC_ME_RUN_PC0 for Periperals, all Periperals are enabled by
default. In order to avoid this problems the Periperals are not enabled
in MC_ME_RUN_PC0, instead we are using MC_ME_RUN_PC1 enabled. The
Periperals which are set for MC_ME_RUN_PC0 are gated. When they are
switch to MC_ME_RUN_PC1, they are ungated.

Issue: ALB-91

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>