s32v234_pcie: Run-time detection of chip revision to ignore ERR009852
commit 58c8283ac070c9fb829a7b0b480e753e62507ec6 from
https://source.codeaurora.org/external/autobsps32/linux

If the chip revision is new enough, PCIe erratum ERR009852 can be safely
ignored regardless of the corresponding Kconfig choice. This is in fact
done by u-boot, which inserts a property in the fdt for us to parse.
If the revision cannot be inferred, fall back to the Kconfig choice.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-876
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>