remoteproc/omap: add a trace to print missing alias ids
commit 11aeec85a5e43a5ecf387ef05a5b80070cef2c3b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The alias ids for OMAP remoteprocs are required by some
rpmsg client drivers to identify a remote processor in
a fixed manner to userspace. Add a trace during probe
to warn developers if the alias id is not defined for a
remoteproc DT node.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>