net: hsr: save timestamp and tag info of stripped skb
commit be083ff10713c5c33619faff36f5d565fee82f08 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Before the tag is stripped from a received skb, the tag info and
receive timestamp is saved in the shared info area of the skb so
that upper core layer can pass the saved info to user application.

In this patch the support of saving tag info and timestamp is added for
HSR PTP packets over L2.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>