net: hsr: save cut-through tx timestamp of stripped skb
commit 78025d556d170cd9192a6d3196b070d68ea9307c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

For a received PTP event message, there may be a cut-through tx timestamp
associated with it also. When the skb is stripped of the HSR tag and a
new skb is created, the cut-through tx timestamp is saved in the shared
redundant info area of the newly created skb.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>