spi: spi-fsl-dspi: Do not modify SPI_RSER or SPI_CTAR while running
commit 0f94259247704f5333281e38e59bc6065833fd5c from
https://source.codeaurora.org/external/autobsps32/linux

Based on commit 7b1eb218fd91 ("fsl-dspi: Not all regs can be modified
when running").

SPI_RSER and SPI_CTAR should be accessed only when DSPI is in stopped
state. The module stops or transitions from Running to Stopped after the
current frame when any one of the following conditions exist:
- SR[EOQF] bit is set
- Chip in the Debug mode and the MCR[FRZ] bit is set
- MCR[HALT] bit is set

Issue: ALB-2511
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Dan Nica <dan.nica@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>