drm/omap: Add ovl checking funcs to dispc_ops
commit 7ce948f595dc4dbdc0768b5a840db39681135233 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In order to be able to dynamically assign overlays to planes we need to
be able to asses the overlay capabilities.

Add a helper function to be able to retrieve the supported capabilities
of an overlay.

And export the function to check if a fourcc is supported on a given
overlay.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>