drm/omap: drop unneeded locking from mgr_fld_write()
commit 8bb04f390b031d0249c6407d4e3ecc727a57832a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Commit d49cd15550d9d4495f6187425318c245d58cb63f ("OMAPDSS: DISPC: lock
access to DISPC_CONTROL & DISPC_CONFIG") added locking to
mgr_fld_write(). This was needed in omapfb times due to lack of good
locking, especially in the case of both V4L2 and fbdev layers using the
DSS driver.

This is not needed for omapdrm, so we can remove the locking.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>