can: flexcan: Keep real format of received frames
commit 17cdaeae689d90ac0ebefbac0344b7bc555848ef from
https://source.codeaurora.org/external/autobsps32/linux

Frames must be fed to the kernel as classic or FD based only on the EDL
bit of the Rx MB, not on whether the interface has been configured with
"fd on". For example, candump from canutils 4.0.6 does not show CAN FD
frames, but should still be able to show the classic ones.

Socket buffers are now allocated in mailbox_read. Although not visible
in the FlexCAN case, they must be also freed from inside mailbox_read if
errors are encountered after allocation.

Issue: ALB-3662
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>