arm64: dts: ti: k3-j721e-som-p0: Reserve memory for IPC between RTOS cores
commit 4667a29ec760557990fb953771a55bfba81fe4a5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a reserved memory node to reserve a portion of the DDR memory to be
used for performing inter-processor communication between all the remote
processors running RTOS on the TI J721E EVM boards. 28 MB of memory is
reserved for this purpose, and this accounts for all the vrings and vring
buffers between all the possible pairs of remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>