arm64: dts: ti: j721e-common-proc-board: Analog audio support
commit 31f1b48365b337739691bae9d681b21ad0267304 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The codec is wired in multi DAI setup (DIN1/2/3/4/DOUT1/2/3 is connected to
McASP serializer).

For some reason with DIN1/DOUT1 setup we can not get 96KHz audio to work as
we would need to run REFCLK2 in 256xFS rate, but then in all sampling rates
the channel 4-7 is going to be muted :o

For other unknown reason 256xFS SCKI does not have this quirk and we can
use 96KHz as well.

There are couple of notable difference compared to DIN1/DOUT1 mode:
the channel mapping is 'random' for first look compared to the single
serializer setup.

Due to the audio setup in hardware with the generic sound card the features
we can support is limited.

With the custom audio driver it is possible to support both 48KHz and
44.1Khz sampling rate families by switching between PLL4/15 as needed.

We can also support both 16 and 24 bit audio in this case as with a custom
driver we can change the slot_width on the bus.

So the channel mapping:
        _      _      _
       |o|c1  |o|p1  |o|p3
 _     | |    | |    | |
|o|c3  |o|c2  |o|p4  |o|p2
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>
------------------------

c1/2/3 - capture jacks (3rd is line)
p1/2/3/4 - playback jacks (4th is line)

2 channel audio (stereo):
0 (left):  p1/c1 left
1 (right): p1/c1 right

4 channel audio:
0: p1/c1 left
1: p2/c2 left
2: p1/c1 right
3: p2/c2 right

6 channel audio
0: p1/c1 left
1: p2/c2 left
2: p3/c3 left
3: p1/c1 right
4: p2/c2 right
5: p3/c3 right

8 channel audio
0: p1/c1 left
1: p2/c2 left
2: p3/c3 left
3: p4 left
4: p1/c1 right
5: p2/c2 right
6: p3/c3 right
7: p4 right

The reason is that we need to put data to all enabled serializers first.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>