can: flexcan: Handle separate interrupt lines
On S32V234, the FlexCAN error and state change interrupts are not
combined with the MB ones. See the S32V234_Interrupt_Routing.xlsx RM
attachment. On S32GEN1, there are separate interrupts for state change,
bus errors, MBs 0-7 and MBs 8-127 respectively.

All the IRQ numbers will be stored in the device private data instead of
the irq field of the net_device structure. That is, the interrupt number
will no longer be featured in ifconfig (as we now have multiple of
them), but we have not supported modifying it anyway.

Issue: ALB-3910
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
[Xulin: revise some codes to apply the patch]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>